fix(sst): fixed sst timing model and updated build files
