+++ lib/common/sst/timing_model/include/timingExpression.h
@@ -1,3 +1,6 @@
+#pragma once
+
+#include <cstdint>
 #include <functional>
 #include <map>
 #include <memory>
@@ -5,12 +8,11 @@
 #include <string>
 #include <vector>
 
-class TimingExpression;
+class TimingState;
 class TimingEvent;
+class TimingOperator;
 class TransitionOperator;
 class RepetitionOperator;
-class TimingState;
-class TimingOperator;
 
 class TimingExpression {
 public:
@@ -20,6 +22,7 @@ public:
   virtual std::string toString() const = 0;
   virtual uint64_t lastEventId() const;
   virtual std::string lastEventName() const;
+  std::shared_ptr<TimingExpression> parent_expression;
 };
 
 class TimingState {
@@ -32,6 +35,7 @@ private:
   std::vector<uint64_t> levels_step;
 
   std::vector<std::shared_ptr<TimingOperator>> operator_queue;
+  std::shared_ptr<TimingExpression> event0;
 
   TimingState &buildRepetition(uint64_t iterations, uint64_t delay);
   TimingState &buildRepetition(uint64_t iterations, uint64_t delay,
@@ -52,7 +56,11 @@ public:
 
   uint64_t getLastScheduledCycle() const;
   void updateLastScheduledCycle(uint64_t cycle);
+  static TimingState createFromEvent(std::shared_ptr<TimingEvent> event);
   static TimingState createFromEvent(const std::string &name);
+  static TimingState
+  createFromExpression(std::shared_ptr<TimingExpression> expression);
+  TimingState &addEvent(std::shared_ptr<TimingEvent> event);
   TimingState &addEvent(const std::string &name, std::function<void()> handler);
   TimingState &addEvent(const std::string &name, uint8_t priority,
                         std::function<void()> handler);
@@ -76,91 +84,8 @@ public:
   void addEventName(const std::string &name);
   std::string toString() const;
 
-  RepetitionOperator getRepetitionOperatorFromLevel(uint64_t level) const;
-};
+  const RepetitionOperator &
+  getRepetitionOperatorFromLevel(uint64_t level) const;
 
-class TimingEvent : public TimingExpression,
-                    public std::enable_shared_from_this<TimingEvent> {
-private:
-  std::string name;
-  uint64_t eventNumber;
-  std::function<void()> handler;
-  uint8_t priority;
-
-public:
-  TimingEvent(const std::string &name, uint64_t eventNumber);
-  uint64_t scheduleEvents(TimingState &state,
-                          uint64_t startCycle) const override;
-  std::string toString() const override;
-  std::string getName() const;
-  uint64_t getEventNumber() const;
-  uint64_t lastEventId() const override;
-  std::string lastEventName() const override;
-  void execute() const;
-  void setHandler(std::function<void()> handler);
-  std::function<void()> getHandler() const;
-
-  void setPriority(uint8_t priority);
-  uint8_t getPriority() const;
-};
-
-class TimingOperator : public TimingExpression {
-public:
-  virtual ~TimingOperator() {}
-};
-
-class TransitionOperator : public TimingOperator {
-private:
-  uint64_t delay;
-  std::string nextEventName;
-  std::function<void()> handler;
-  std::shared_ptr<TimingExpression> from;
-  std::shared_ptr<TimingExpression> to;
-
-public:
-  TransitionOperator(uint64_t delay, std::string nextEventName,
-                     std::function<void()> handler,
-                     std::shared_ptr<TimingExpression> from,
-                     std::shared_ptr<TimingExpression> to);
-
-  uint64_t getDelay() const;
-  std::string getNextEventName() const;
-  std::function<void()> getHandler() const;
-  std::shared_ptr<TimingExpression> getFrom() const;
-  std::shared_ptr<TimingExpression> getTo() const;
-
-  uint64_t scheduleEvents(TimingState &state,
-                          uint64_t startCycle) const override;
-  std::string toString() const override;
-  uint64_t lastEventId() const override;
-  std::string lastEventName() const override;
-};
-
-class RepetitionOperator : public TimingOperator {
-private:
-  uint64_t iterations;
-  uint64_t delay;
-  uint64_t level;
-  uint64_t step;
-  std::shared_ptr<TimingExpression> expression;
-
-public:
-  RepetitionOperator(uint64_t iterations, uint64_t delay, uint64_t level,
-                     uint64_t step,
-                     std::shared_ptr<TimingExpression> expression);
-
-  uint64_t getIterations() const;
-  uint64_t getDelay() const;
-  uint64_t getLevel() const;
-  uint64_t getStep() const;
-  void setIterations(uint64_t iterations) { this->iterations = iterations; }
-  void setDelay(uint64_t delay) { this->delay = delay; }
-  void setStep(uint64_t step) { this->step = step; }
-  std::shared_ptr<TimingExpression> getExpression() const;
-
-  uint64_t scheduleEvents(TimingState &state,
-                          uint64_t startCycle) const override;
-  std::string toString() const override;
-  uint64_t lastEventId() const override;
-  std::string lastEventName() const override;
+  void print(std::ostream &os);
 };
