+++ lib/resources/dpu/sst/dpu.cpp
@@ -1,6 +1,7 @@
 #include "dpu.h"
 #include "dataEvent.h"
 #include "dpu_operations.h"
+#include "timingOperators.h"
 
 using namespace SST;
