+++ lib/resources/dpu_2cycle_mac/sst/dpu_2cycle_mac.cpp
@@ -1,6 +1,7 @@
 #include "dpu_2cycle_mac.h"
 #include "dataEvent.h"
 #include "dpu_operations.h"
+#include "timingOperators.h"
 
 using namespace SST;
