+++ lib/resources/iosram_both/sst/iosram_both.cpp
@@ -3,6 +3,7 @@
 #include "dataEvent.h"
 #include "ioEvents.h"
 #include "iosram_both_pkg.h"
+#include "timingOperators.h"
 
 using namespace SST;
